requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
cryptography
python-dotenv
pandas
seaborn
matplotlib
faker